requirements.txt
requests
beautifulsoup4
gunicorn
django-heroku
python-telegram-bot
python-telegram-bot-raw